added the tools bar to the Info and Crm fragments;
