+++ app/src/main/java/com/team/mamba/atlas/userInterface/dashBoard/info/InfoNavigator.java
@@ -34,4 +34,10 @@ public interface InfoNavigator {
 
     void restartApplication();
 
+    void onCrmClicked();
+
+    void onNotificationsClicked();
+
+    void onContactsClicked();
+
 }
