+++ app/src/main/res/layouts/dashBoard/crm/layout/crm_layout.xml
@@ -144,6 +144,98 @@
 
             </android.support.v4.widget.SwipeRefreshLayout>
 
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal"
+                android:paddingLeft="25dp"
+                android:paddingRight="25dp">
+
+                <FrameLayout
+                    android:layout_width="55dp"
+                    android:layout_height="55dp">
+
+                    <ImageView
+                        style="@style/IconImageView"
+                        android:background="?attr/selectableItemBackgroundBorderless"
+                        android:src="@drawable/ic_tachometer_solid"
+                        android:layout_gravity="center"
+                        android:tint="@color/material_icon"
+                        android:onClick="@{()-> viewModel.onInfoClicked()}"
+                        />
+
+                </FrameLayout>
+
+                <View
+                    android:layout_width="0dp"
+                    android:layout_height="1dp"
+                    android:layout_weight="1" />
+
+                <FrameLayout
+                    android:layout_width="55dp"
+                    android:layout_height="55dp">
+
+
+
+                    <ImageView
+                        android:id="@+id/iv_contacts"
+                        style="@style/IconImageView"
+                        android:background="?attr/selectableItemBackgroundBorderless"
+                        android:src="@drawable/ic_address_book_solid"
+                        android:onClick="@{()-> viewModel.onContactsClicked()}"
+                        android:layout_gravity="center"
+                        android:tint="@color/material_icon"
+                        />
+
+
+                </FrameLayout>
+
+
+                <View
+                    android:layout_width="0dp"
+                    android:layout_height="1dp"
+                    android:layout_weight="1" />
+
+                <FrameLayout
+                    android:layout_width="55dp"
+                    android:layout_height="55dp">
+
+
+                    <ImageView
+                        android:id="@+id/iv_crm"
+                        style="@style/IconImageView"
+                        android:background="?attr/selectableItemBackgroundBorderless"
+                        android:src="@drawable/ic_pie_chart"
+                        android:layout_gravity="center"
+                        />
+
+                </FrameLayout>
+
+                <View
+                    android:layout_width="0dp"
+                    android:layout_height="1dp"
+                    android:layout_weight="1" />
+
+                <FrameLayout
+                    android:layout_width="55dp"
+                    android:layout_height="55dp">
+
+
+                    <ImageView
+                        android:id="@+id/iv_notifications"
+                        style="@style/IconImageView"
+                        android:background="?attr/selectableItemBackgroundBorderless"
+                        android:src="@drawable/ic_bell_solid"
+                        android:tint="@color/material_icon"
+                        android:onClick="@{()-> viewModel.onNotificationsClicked()}"
+                        android:layout_gravity="center"
+                        />
+
+                </FrameLayout>
+
+            </LinearLayout>
+
+
         </LinearLayout>
 
         <FrameLayout
