+++ app/src/main/res/layouts/dashBoard/layout/fragment_container.xml
@@ -35,134 +35,6 @@
 
             </FrameLayout>
 
-            <LinearLayout
-                android:id="@+id/layout_tool_bar"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:paddingLeft="25dp"
-                android:paddingRight="25dp">
-
-                <FrameLayout
-                    android:layout_width="55dp"
-                    android:layout_height="55dp">
-
-                    <ImageView
-                        android:id="@+id/iv_info_selected"
-                        style="@style/IconImageView"
-                        android:background="?attr/selectableItemBackgroundBorderless"
-                        android:src="@drawable/ic_tachometer_solid"
-                        android:layout_gravity="center"
-                        />
-
-                    <ImageView
-                        android:id="@+id/iv_info_not_selected"
-                        style="@style/IconImageView"
-                        android:background="?attr/selectableItemBackgroundBorderless"
-                        android:src="@drawable/ic_tachometer_solid"
-                        android:tint="@color/grey"
-                        android:onClick="@{()-> viewModel.onInfoClicked()}"
-                        android:visibility="gone"
-                        android:layout_gravity="center"
-                        />
-
-
-                </FrameLayout>
-
-                <View
-                    android:layout_width="0dp"
-                    android:layout_height="1dp"
-                    android:layout_weight="1" />
-
-                <FrameLayout
-                    android:layout_width="55dp"
-                    android:layout_height="55dp">
-
-                    <ImageView
-                        android:id="@+id/iv_contacts_selected"
-                        style="@style/IconImageView"
-                        android:background="?attr/selectableItemBackgroundBorderless"
-                        android:src="@drawable/ic_address_book_solid"
-                        android:visibility="gone"
-                        android:layout_gravity="center"
-                        />
-
-                    <ImageView
-                        android:id="@+id/iv_contacts_not_selected"
-                        style="@style/IconImageView"
-                        android:background="?attr/selectableItemBackgroundBorderless"
-                        android:src="@drawable/ic_address_book_solid"
-                        android:tint="@color/material_icon"
-                        android:onClick="@{()-> viewModel.onContactsClicked()}"
-                        android:layout_gravity="center"
-                        />
-
-
-                </FrameLayout>
-
-
-                <View
-                    android:layout_width="0dp"
-                    android:layout_height="1dp"
-                    android:layout_weight="1" />
-
-                <FrameLayout
-                    android:layout_width="55dp"
-                    android:layout_height="55dp">
-
-
-                <ImageView
-                    android:id="@+id/iv_crm_selected"
-                    style="@style/IconImageView"
-                    android:background="?attr/selectableItemBackgroundBorderless"
-                    android:src="@drawable/ic_pie_chart"
-                    android:visibility="gone"
-                    android:layout_gravity="center"
-                    />
-
-                <ImageView
-                    android:id="@+id/iv_crm_not_selected"
-                    style="@style/IconImageView"
-                    android:background="?attr/selectableItemBackgroundBorderless"
-                    android:src="@drawable/ic_pie_chart"
-                    android:tint="@color/material_icon"
-                    android:onClick="@{()-> viewModel.onCrmClicked()}"
-                    android:layout_gravity="center"
-                    />
-
-                </FrameLayout>
-
-                <View
-                    android:layout_width="0dp"
-                    android:layout_height="1dp"
-                    android:layout_weight="1" />
-
-                <FrameLayout
-                    android:layout_width="55dp"
-                    android:layout_height="55dp">
-
-                <ImageView
-                    android:id="@+id/iv_notifications_selected"
-                    style="@style/IconImageView"
-                    android:background="?attr/selectableItemBackgroundBorderless"
-                    android:src="@drawable/ic_bell_solid"
-                    android:visibility="gone"
-                    android:layout_gravity="center"
-                    />
-
-                <ImageView
-                    android:id="@+id/iv_notifications_not_selected"
-                    style="@style/IconImageView"
-                    android:background="?attr/selectableItemBackgroundBorderless"
-                    android:src="@drawable/ic_bell_solid"
-                    android:tint="@color/material_icon"
-                    android:onClick="@{()-> viewModel.onNotificationsClicked()}"
-                    android:layout_gravity="center"
-                    />
-
-                </FrameLayout>
-
-            </LinearLayout>
 
         </LinearLayout>
